% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excercise2.R
\name{calc_something_big}
\alias{calc_something_big}
\title{Imagine this is a resource-intensive calculation}
\usage{
calc_something_big()
}
\value{
results as a dataframe
}
\description{
Also imagine that it is always correct and you don't need to test it
}
