% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excercise2.R
\name{do_something}
\alias{do_something}
\title{Imagine this does something useful}
\usage{
do_something(some_Var = some_var)
}
\arguments{
\item{some_Var}{What to insert into the dataframe}
}
\value{
a dataframe with results
}
\description{
Imagine this does something useful
}
