% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whale.R
\name{say}
\alias{say}
\title{Whale Echoes Phrase}
\usage{
say(what)
}
\arguments{
\item{what}{This is the phrase you wish the whale to repeat. A random phrase
will be used if nothing is entered.}
}
\value{
Whale shaped message.
}
\description{
Whale Echoes Phrase
}
\examples{
say("Hi I'm a Whale")
}
